Remove confusing dupliate method for specifying pp components. (#25)
We had a list of components to do, and then define the components
in the yaml. This is confusing and results in cylc graph problems
with a component is listed but not defined.
Instead, set a switch in the component definitions, as a signal
to do or not to do.

Co-authored-by: Chris Blanton <Chris.Blanton@ldt-4359643.gfdl.noaa.gov>